requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
attrs==19.3.0
bleach==1.5.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.2
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
html5lib==0.9999999
importlib-metadata==1.5.0
kiwisolver==1.1.0
Markdown==3.2.1
more-itertools==8.2.0
networkx==2.2
numpy==1.18.1
opencv-python==4.1.2.30
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyglet==1.4.10
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
Tensorforce==0.5.3
termcolor==1.1.0
tqdm==4.42.1
trfl==1.0.2
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==2.2.0
